% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interaction_breakdown.R
\name{get_interaction_breakdown}
\alias{get_interaction_breakdown}
\title{get_interaction_breakdown}
\usage{
get_interaction_breakdown(
  info,
  type = c("ss", "jn"),
  p_adjust = "fdr",
  a_thr = 0.05
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
get_interaction_breakdown
}
