% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jn_plot.R
\name{get_jn_plot}
\alias{get_jn_plot}
\title{get_jn_plot}
\usage{
get_jn_plot(
  model,
  x_var,
  m_vars = NULL,
  facet_wrap_opts = list(ncol = 3, scales = "fixed", labeller = "label_both")
)
}
\arguments{
\item{m_vars}{}
}
\value{

}
\description{
get_jn_plot
}
\concept{viz}
