% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_info.R
\name{get_model_info}
\alias{get_model_info}
\title{get_model_info}
\usage{
get_model_info(model, x_var = NULL, m_vars = NULL, x_step = 0.01, round = 3)
}
\arguments{
\item{round}{}
}
\value{

}
\description{
get_model_info
}
