% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_scale_from_formula}
\alias{remove_scale_from_formula}
\title{remove_scale_from_formula}
\usage{
remove_scale_from_formula(formula, term)
}
\arguments{
\item{term}{}
}
\value{

}
\description{
remove_scale_from_formula
}
\concept{util}
