% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_remove_fun_from_term}
\alias{str_remove_fun_from_term}
\title{str_remove_fun_from_term}
\usage{
str_remove_fun_from_term(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
str_remove_fun_from_term
}
\concept{util}
