% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_remove_scale_from_formula}
\alias{str_remove_scale_from_formula}
\title{str_remove_scale_from_formula}
\usage{
str_remove_scale_from_formula(formula, term)
}
\arguments{
\item{term}{}
}
\value{

}
\description{
str_remove_scale_from_formula
}
\concept{util}
